make generator bin and lib with custom modules
