+++ crates/generator/Cargo.toml
@@ -1,9 +1,7 @@
 [package]
 name = "e3-greco-generator"
-description = "Generator of cryptographic parameters and constants for Greco zk-proofs"
 version.workspace = true
 edition.workspace = true
-license = "MIT"
 
 [lib]
 name = "greco_generator"
@@ -11,20 +9,20 @@ path = "src/lib.rs"
 
 [[bin]]
 name = "generator"
-path = "src/generator/main.rs"
+path = "src/bin/generator.rs"
 
 [dependencies]
 polynomial = { workspace = true }
+blake3 = { workspace = true }
 fhe = { workspace = true }
-fhe-traits = { workspace = true }
 fhe-math = { workspace = true }
-serde_json = { workspace = true }
-rand = { workspace = true }
-serde = { workspace = true }
+fhe-traits = { workspace = true }
+itertools = { workspace = true }
 num-bigint = { workspace = true }
 num-traits = { workspace = true }
-ndarray = { workspace = true }
-itertools = { workspace = true }
+rand = { workspace = true }
 rayon = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
 toml = { workspace = true }
-blake3 = { workspace = true }
+clap = { version = "4.0", features = ["derive"] }
