+++ crates/generator/src/generators/mod.rs
@@ -1,0 +1,7 @@
+//! Output generators for different formats.
+//!
+//! This module contains generators for various output formats including
+//! Noir constants and Prover TOML files.
+
+pub mod noir;
+pub mod toml;
